+++ CHANGELOG.md
@@ -1,4 +1,6 @@
 # Changelog
+## Unreleased
+- Add new version of word2vec analyzer
 ## [Version 0.2.13](https://github.com/NVISO-BE/ee-outliers/releases/tag/0.2.13) (April 8, 2020)
 - Improved documentation (source code and user documentation)
 - Fixes an issue where DSL queries in use case configuration files would not be correctly parsed (issue #455)